removing layout and statuslayout and addin applayout
