+++ apps/financial-aid/web-osk/src/components/HasApplied/hasApplied.tsx
@@ -3,6 +3,7 @@ import { Text, Button } from '@island.is/island-ui/core'
 
 import { UserContext } from '@island.is/financial-aid-web/osk/src/components/UserProvider/UserProvider'
 
+import { ContentContainer } from '@island.is/financial-aid-web/osk/src/components'
 import { useRouter } from 'next/router'
 import { Routes } from '@island.is/financial-aid/shared/lib'
 
@@ -11,7 +12,7 @@ const HasApplied = () => {
   const router = useRouter()
 
   return (
-    <>
+    <ContentContainer>
       <Text as="h1" variant="h2" marginBottom={[3, 3, 5]}>
         Abbabbab {user?.name ?? ''} <br />
         þú hefur þegar sótt um fyrir þennan mánuð
@@ -26,7 +27,7 @@ const HasApplied = () => {
       >
         Kíktu á stöðu síðuna{' '}
       </Button>
-    </>
+    </ContentContainer>
   )
 }
