+++ apps/financial-aid/web-osk/src/components/SideBar/SideBar.tsx
@@ -1,0 +1,26 @@
+import React, { useContext, useEffect } from 'react'
+import { Box, Text } from '@island.is/island-ui/core'
+
+import { Stepper } from '@island.is/financial-aid-web/osk/src/components'
+
+import { UserContext } from '@island.is/financial-aid-web/osk/src/components/UserProvider/UserProvider'
+
+const SideBar = () => {
+  const { user } = useContext(UserContext)
+
+  return (
+    <>
+      {user?.currentApplication ? (
+        <Box>
+          <Text as="h3" variant="h3" marginBottom={[1, 1, 2]}>
+            Umsókn um fjárhagsaðstoð hjá Hafnarfjarðarbæ
+          </Text>
+        </Box>
+      ) : (
+        <Stepper />
+      )}
+    </>
+  )
+}
+
+export default SideBar
