+++ apps/financial-aid/web-osk/src/components/index.ts
@@ -2,9 +2,6 @@ export { default as Header } from './Header/Header'
 export { default as FormProvider } from './FormProvider/FormProvider'
 export { default as UserProvider } from './UserProvider/UserProvider'
 export { default as MunicipalityProvider } from './MunicipalityProvider/MunicipalityProvider'
-export { default as FormLayout } from './FormLayout/FormLayout'
-export { default as Layout } from './Layout/Layout'
-export { default as StatusLayout } from './StatusLayout/StatusLayout'
 export { default as Footer } from './Footer/Footer'
 export { default as ContentContainer } from './ContentContainer/ContentContainer'
 export { default as Logo } from './Logo/index'
@@ -25,3 +22,6 @@ export { default as UserInfo } from './Summary/UserInfo'
 export { default as AllFiles } from './Summary/AllFiles'
 export { default as FormInfo } from './Summary/FormInfo'
 export { default as FormComment } from './Summary/FormComment'
+export { default as AppLayout } from './AppLayout/AppLayout'
+export { default as Stepper } from './Stepper/Stepper'
+export { default as SideBar } from './SideBar/SideBar'
