+++ apps/financial-aid/web-osk/src/routes/application/AddressForm/addressForm.tsx
@@ -4,7 +4,6 @@ import { Text, RadioButton, Input, Box } from '@island.is/island-ui/core'
 import {
   ContentContainer,
   Footer,
-  Layout,
   RadioButtonContainer,
 } from '@island.is/financial-aid-web/osk/src/components'
 
@@ -41,7 +40,7 @@ const AddressForm = () => {
   ]
 
   return (
-    <Layout>
+    <>
       <ContentContainer>
         <Text as="h1" variant="h2" marginBottom={[3, 3, 4]}>
           Hvar býrðu?
@@ -137,7 +136,7 @@ const AddressForm = () => {
           }
         }}
       />
-    </Layout>
+    </>
   )
 }
