+++ apps/financial-aid/web-osk/src/routes/application/BankInfoForm/bankInfoForm.tsx
@@ -4,7 +4,6 @@ import { Text, Input, Box } from '@island.is/island-ui/core'
 import {
   ContentContainer,
   Footer,
-  Layout,
 } from '@island.is/financial-aid-web/osk/src/components'
 import { FormContext } from '@island.is/financial-aid-web/osk/src/components/FormProvider/FormProvider'
 
@@ -103,7 +102,7 @@ const Form = () => {
   }
 
   return (
-    <Layout>
+    <>
       <ContentContainer>
         <Text as="h1" variant="h2" marginBottom={2}>
           Greiðsla fjárhagsaðstoðar
@@ -162,7 +161,7 @@ const Form = () => {
             : 'Gefa upp seinna'
         }
       />
-    </Layout>
+    </>
   )
 }
