+++ apps/financial-aid/web-osk/src/routes/application/PersonalTaxCreditForm/personalTaxCreditForm.tsx
@@ -4,7 +4,6 @@ import { Text } from '@island.is/island-ui/core'
 import {
   ContentContainer,
   Footer,
-  Layout,
   RadioButtonContainer,
 } from '@island.is/financial-aid-web/osk/src/components'
 
@@ -50,7 +49,7 @@ const PersonalTaxCreditForm = () => {
   }
 
   return (
-    <Layout>
+    <>
       <ContentContainer>
         <Text as="h1" variant="h2" marginBottom={2}>
           Viltu nota persónuafslátt?
@@ -109,7 +108,7 @@ const PersonalTaxCreditForm = () => {
         previousUrl={navigation?.prevUrl}
         onNextButtonClick={() => errorCheck()}
       />
-    </Layout>
+    </>
   )
 }
