+++ apps/financial-aid/web-osk/src/routes/application/SummaryForm/summaryForm.tsx
@@ -4,7 +4,6 @@ import { Text, Divider, Box } from '@island.is/island-ui/core'
 import {
   ContentContainer,
   Footer,
-  Layout,
   CancelModal,
   Estimation,
   UserInfo,
@@ -121,7 +120,7 @@ const SummaryForm = () => {
   ) as NavigationProps
 
   return (
-    <Layout>
+    <>
       <ContentContainer>
         <Text as="h1" variant="h2" marginBottom={[3, 3, 4]}>
           Yfirlit umsóknar
@@ -184,7 +183,7 @@ const SummaryForm = () => {
         nextButtonText="Senda umsókn"
         onNextButtonClick={handleNextButtonClick}
       />
-    </Layout>
+    </>
   )
 }
