+++ apps/financial-aid/web-osk/src/routes/application/TaxReturnForm/taxReturnForm.tsx
@@ -4,7 +4,6 @@ import { Text, LinkContext } from '@island.is/island-ui/core'
 import {
   ContentContainer,
   Footer,
-  Layout,
   Files,
 } from '@island.is/financial-aid-web/osk/src/components'
 import { FormContext } from '@island.is/financial-aid-web/osk/src/components/FormProvider/FormProvider'
@@ -30,7 +29,7 @@ const TaxReturnForm = () => {
   }
 
   return (
-    <Layout>
+    <>
       <ContentContainer>
         <Text as="h1" variant="h2" marginBottom={2}>
           Skattframtal
@@ -83,7 +82,7 @@ const TaxReturnForm = () => {
         nextButtonText="Halda áfram"
         onNextButtonClick={() => errorCheck()}
       />
-    </Layout>
+    </>
   )
 }
